+++ src/api/AxiosClient.ts
@@ -1,6 +1,7 @@
-import axios, { AxiosError } from 'axios';
+import axios from 'axios';
 import type { AxiosInstance } from 'axios';
 import { API_URLS, BASE_URL } from '../constant/config';
+import { refreshAccessAPI } from './authClient';
 
 const axiosClient: AxiosInstance = axios.create({
   baseURL: BASE_URL,
@@ -33,42 +34,6 @@ axiosClient.interceptors.request.use(
   (error) => Promise.reject(error)
 );
 
-/**
- * - access_token을 갱신하는 함수
- * - 함수 아래 interceptors만 의존해야 하기 때문에 export하지 않음
- * - API 명세의 요구에 맞춰 header를 활용하고 refresh token을 설정
- */
-async function refreshAccessAPI() {
-  try {
-    const sessionToken = sessionStorage.getItem('sessionToken');
-    if (!sessionToken) throw Error('sessionToken');
-
-    const {
-      data: { access_token },
-    } = await authClient.post<{
-      success: boolean;
-      access_token: string;
-    }>(API_URLS.REFRESH, null, {
-      headers: {
-        Authorization: `Bearer ${sessionToken.slice(
-          1,
-          sessionToken.length - 1
-        )}`,
-      },
-    });
-
-    localStorage.setItem('accessToken', `"${access_token}"`);
-
-    return access_token;
-  } catch (error) {
-    localStorage.clear();
-    sessionStorage.clear();
-    if (error instanceof AxiosError) {
-      return error.response?.data;
-    }
-  }
-}
-
 axiosClient.interceptors.response.use(
   (res) => res,
   async (err) => {
@@ -76,6 +41,7 @@ axiosClient.interceptors.response.use(
       config,
       response: { status },
     } = err;
+
     if (config.url === API_URLS.REFRESH || status !== 401 || config.sent) {
       return Promise.reject(err);
     }
@@ -85,6 +51,7 @@ axiosClient.interceptors.response.use(
     if (accessToken) {
       config.headers.Authorization = `Bearer ${accessToken}`;
     }
+
     return axiosClient(config);
   }
 );
